Yenesis-main/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
beautifulsoup4==4.13.4
boto3==1.38.16
botocore==1.38.16
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
dotenv==0.9.9
fastjsonschema==2.21.1
finnhub-python==2.4.23
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.6.0
fyers_apiv3==3.1.7
idna==3.10
jmespath==0.10.0
kiwisolver==1.4.8
logger==1.4
matplotlib==3.10.1
mplfinance==0.12.10b0
multidict==6.4.3
multitasking==0.0.11
numpy==1.23.5
packaging==25.0
pandas==2.2.3
pandas-ta==0.3.14b0
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.3
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.31.0
s3transfer==0.12.0
six==1.17.0
soupsieve==2.7
ta==0.11.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.6.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.61
mysql-connector-python==8.0.33
SQLAlchemy==2.0.23
